[Feature] #46 - API DTO 파일 통합
