--- Growthook/Growthook/Data/DTO/ActionListDTO/ActionListPercentResponse.swift
@@ -1,17 +1,0 @@
-//
-//  ActionListPercentResponse.swift
-//  Growthook
-//
-//  Created by 천성우 on 1/7/24.
-//
-
-import Foundation
-
-struct ActionListPercentResponse: Codable {
-    let data: Int
-    
-    func convertToActionListPersentModel() -> ActionListPersentModel {
-        return ActionListPersentModel(actionListPercent: data)
-    }
-}
-
