--- Growthook/Growthook/Data/Model/ActionListModel/ActionListDoingModel.swift
@@ -1,15 +1,0 @@
-//
-//  ActionListDoingModel.swift
-//  Growthook
-//
-//  Created by 천성우 on 1/8/24.
-//
-
-import Foundation
-
-struct ActionListDoingModel {
-    let actionPlanID: Int
-    let content: String
-    let isScraped: Bool
-    let seedID: Int
-}
