--- Growthook/Growthook/Data/Model/ActionListModel/ActionListPercentModel.swift
@@ -1,12 +1,0 @@
-//
-//  ActionListPercentModel.swift
-//  Growthook
-//
-//  Created by 천성우 on 1/7/24.
-//
-
-import Foundation
-
-struct ActionListPersentModel {
-    let actionListPercent: Int
-}
